refactor: Item 과 PopularItem 서비스 및 레포지토리 분리
