+++ src/main/java/kr/hhplus/be/server/application/item/ItemFacadeService.java
@@ -1,20 +1,28 @@
 package kr.hhplus.be.server.application.item;
 
 import kr.hhplus.be.server.domain.item.ItemService;
+import kr.hhplus.be.server.domain.item.PopularItem;
+import kr.hhplus.be.server.domain.item.PopularItemDetail;
+import kr.hhplus.be.server.domain.item.PopularItemService;
 import kr.hhplus.be.server.domain.order.OrderItemStatistics;
 import kr.hhplus.be.server.domain.order.OrderService;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+import java.util.Optional;
+
 @Service
 public class ItemFacadeService {
 
     private final ItemService itemService;
+    private final PopularItemService popularItemService;
     private final OrderService orderService;
 
-    public ItemFacadeService(ItemService itemService, OrderService orderService) {
+    public ItemFacadeService(ItemService itemService, PopularItemService popularItemService, OrderService orderService) {
         this.itemService = itemService;
+        this.popularItemService = popularItemService;
         this.orderService = orderService;
     }
 
@@ -24,6 +32,12 @@ public class ItemFacadeService {
 
         OrderItemStatistics orderItemStatistics = orderService.findYesterdayOrderItemStatistics();
 
-        itemService.createPopularItems(orderItemStatistics);
+        popularItemService.createPopularItems(orderItemStatistics);
+    }
+
+    public List<PopularItemDetail> findPopularItemDetails() {
+        return popularItemService.findPopularItems().stream()
+                .map(p -> PopularItemDetail.of(p, itemService.findById(p.itemId())))  // ← 프록시 경유
+                .toList();
     }
 }
