+++ src/main/java/kr/hhplus/be/server/domain/item/ItemRepository.java
@@ -2,7 +2,6 @@ package kr.hhplus.be.server.domain.item;
 
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
@@ -10,11 +9,7 @@ public interface ItemRepository {
 
     Optional<Item> findById(long id);
 
-    List<PopularItem> findPopularItems();
-
     Item findByIdWithLock(long id);
-
-    List<PopularItemStatistics> savePopularItems(List<PopularItemStatistics> popularItemStatistics);
 
     Item saveItem(Item item);
 }
