+++ src/main/java/kr/hhplus/be/server/domain/item/ItemService.java
@@ -1,13 +1,9 @@
 package kr.hhplus.be.server.domain.item;
 
-import kr.hhplus.be.server.domain.order.OrderItemStatistics;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 import static kr.hhplus.be.server.domain.item.ItemCommand.*;
 
@@ -25,15 +21,6 @@ public class ItemService {
         return itemRepository.findById(id).orElseThrow(() -> new RuntimeException("상품을 찾을 수 없습니다."));
     }
 
-    public List<PopularItemDetail> findPopularItems() {
-
-        List<PopularItem> popularItems = Optional.ofNullable(itemRepository.findPopularItems()).orElse(List.of());
-
-        return popularItems.stream()
-                .map(popularItem -> PopularItemDetail.of(popularItem, findById(popularItem.itemId())))
-                .toList();
-    }
-
     @Transactional
     public Item decreaseStock(StockDecreaseCommand command) {
 
@@ -41,16 +28,6 @@ public class ItemService {
         item.decreaseStock(command.count());
 
         return item;
-    }
-
-    @Transactional
-    public List<PopularItemStatistics> createPopularItems(OrderItemStatistics orderItemStatistics) {
-
-        List<PopularItemStatistics> popularItemStatistics = orderItemStatistics.getItemIds().stream()
-                .map(itemId -> PopularItemStatistics.of(itemId, orderItemStatistics.getOrderDate(itemId), orderItemStatistics.getTotalOrderCount(itemId)))
-                .toList();
-
-        return itemRepository.savePopularItems(popularItemStatistics);
     }
 
     @CacheEvict(value = "cache:item", key = "#itemId")
