+++ src/test/java/kr/hhplus/be/server/infrastructure/support/RedisCleanup.java
@@ -1,0 +1,18 @@
+package kr.hhplus.be.server.infrastructure.support;
+
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.stereotype.Component;
+
+@Component
+public class RedisCleanup {
+
+    private final RedisConnectionFactory factory;
+
+    public RedisCleanup(RedisConnectionFactory factory) {
+        this.factory = factory;
+    }
+
+    public void flushAll() {
+        factory.getConnection().flushAll();
+    }
+}
